Combine Lumberjack tests into serial and mpi test executables
