+++ scripts/llnl_scripts/llnl_lc_build_tools.py
@@ -290,7 +290,7 @@ def uberenv_build(prefix, spec, project_file, config_dir, mirror_path):
     # Move files generated by spack in source directory to TPL install directory
     print "[Moving spack generated files to TPL build directory]"
     repo_dir = get_repo_dir()
-    for file in ["spack-build-env.txt", "spack-build-out.txt"]:
+    for file in ["spack-build-env.txt", "spack-build-out.txt", "spack-configure-args.txt"]:
         src = pjoin(repo_dir, file)
         dst = pjoin(prefix, "{0}-{1}".format(spec.replace(" ", "_"),file))
         if os.path.exists(src) and not os.path.exists(dst):
