+++ src/axom/lumberjack/tests/CMakeLists.txt
@@ -7,65 +7,62 @@
 #------------------------------------------------------------------------------
 
 #------------------------------------------------------------------------------
-# Specify list of tests
+# Add serial GTests based tests
 #------------------------------------------------------------------------------
+set(lumberjack_serial_tests
+    lumberjack_Lumberjack.hpp
+    lumberjack_Message.hpp
+    lumberjack_TextEqualityCombiner.hpp )
 
-# Add Serial GTests based tests
-set(gtest_lumberjack_tests
-    lumberjack_Lumberjack.cpp
-    lumberjack_Message.cpp
-    lumberjack_TextEqualityCombiner.cpp
-    )
+blt_add_executable(NAME       lumberjack_serial_tests
+                   SOURCES    lumberjack_serial_main.cpp ${lumberjack_serial_tests}
+                   OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
+                   DEPENDS_ON lumberjack gtest 
+                   FOLDER     axom/lumberjack/tests )
 
-foreach(test ${gtest_lumberjack_tests})
-    get_filename_component( test_name ${test} NAME_WE )
-    blt_add_executable( NAME ${test_name}_test
-                        SOURCES ${test}
-                        OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                        DEPENDS_ON lumberjack gtest 
-                        FOLDER axom/lumberjack/tests )
-    blt_add_test( NAME ${test_name} 
-                  COMMAND ${test_name}_test )
+foreach(test_suite ${lumberjack_serial_tests})
+    get_filename_component(test_suite ${test_suite} NAME_WE)
+    blt_add_test(NAME    ${test_suite}
+                 COMMAND lumberjack_serial_tests --gtest_filter=${test_suite}* )
 endforeach()
 
 
 #------------------------------------------------------------------------------
 # Add MPI GTest based tests
 #------------------------------------------------------------------------------
-set(mpi_gtest_lumberjack_tests
-    lumberjack_BinaryCommunicator.cpp
-    lumberjack_RootCommunicator.cpp
-    )
+set(lumberjack_mpi_tests
+    lumberjack_BinaryCommunicator.hpp
+    lumberjack_RootCommunicator.hpp )
 
-foreach(test ${mpi_gtest_lumberjack_tests})
-    get_filename_component( test_name ${test} NAME_WE )
-    blt_add_executable( NAME ${test_name}_test
-                        SOURCES ${test}
-                        OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                        DEPENDS_ON lumberjack gtest 
-                        FOLDER axom/lumberjack/tests )
+blt_add_executable(NAME       lumberjack_mpi_tests
+                   SOURCES    lumberjack_mpi_main.cpp ${lumberjack_mpi_tests}
+                   OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
+                   DEPENDS_ON lumberjack mpi gtest 
+                   FOLDER     axom/lumberjack/tests )
+
+foreach(test_suite ${lumberjack_mpi_tests})
+    get_filename_component( test_suite ${test_suite} NAME_WE )
     foreach(task_count RANGE 1 8)
-        blt_add_test( NAME ${test_name}_${task_count}
-                      COMMAND ${test_name}_test
-                      NUM_MPI_TASKS ${task_count})
+        blt_add_test(NAME          ${test_suite}_${task_count}
+                     COMMAND       lumberjack_mpi_tests --gtest_filter=${test_suite}*
+                     NUM_MPI_TASKS ${task_count} )
     endforeach()
 endforeach()
 
 #------------------------------------------------------------------------------
 # Single source file tests
 #------------------------------------------------------------------------------
-blt_add_executable(NAME lumberjack_speed_test
-                   SOURCES lumberjack_speedTest.cpp
+blt_add_executable(NAME       lumberjack_speed_test
+                   SOURCES    lumberjack_speedTest.cpp
                    OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                   DEPENDS_ON lumberjack
-                   FOLDER axom/lumberjack/tests )
+                   DEPENDS_ON lumberjack mpi
+                   FOLDER     axom/lumberjack/tests )
 
 set(lumberjack_sample_input_dir ${PROJECT_SOURCE_DIR}/axom/lumberjack/tests/sampleInput)
-blt_add_test( NAME lumberjack_speedTest_binary
-              COMMAND lumberjack_speed_test b 10 ${lumberjack_sample_input_dir}/loremIpsum02
-              NUM_MPI_TASKS 4)
+blt_add_test(NAME          lumberjack_speedTest_binary
+             COMMAND       lumberjack_speed_test b 10 ${lumberjack_sample_input_dir}/loremIpsum02
+             NUM_MPI_TASKS 4)
               
-blt_add_test( NAME lumberjack_speedTest_root
-              COMMAND lumberjack_speed_test r 10 ${lumberjack_sample_input_dir}/loremIpsum02
-              NUM_MPI_TASKS 4)
-              
+blt_add_test(NAME          lumberjack_speedTest_root
+             COMMAND       lumberjack_speed_test r 10 ${lumberjack_sample_input_dir}/loremIpsum02
+             NUM_MPI_TASKS 4)
