+++ src/axom/lumberjack/tests/lumberjack_RootCommunicator.hpp
@@ -11,17 +11,15 @@
 
 #include "mpi.h"
 
-#include "axom/lumberjack/BinaryTreeCommunicator.hpp"
-
-#include "axom/core/utilities/Utilities.hpp"
+#include "axom/lumberjack/RootCommunicator.hpp"
 
 
-TEST(lumberjack_BinaryCommunicator, basic)
+TEST(lumberjack_RootCommunicator, basic)
 {
   MPI_Barrier(MPI_COMM_WORLD);
 
   const int ranksLimit = 5;
-  axom::lumberjack::BinaryTreeCommunicator c;
+  axom::lumberjack::RootCommunicator c;
   c.initialize(MPI_COMM_WORLD, ranksLimit);
 
   int commRank = -1;
@@ -49,7 +47,7 @@ TEST(lumberjack_BinaryCommunicator, basic)
   EXPECT_EQ(c.ranksLimit(), newRanksLimit);
 
   // Check number of pushes
-  EXPECT_EQ(c.numPushesToFlush(), axom::utilities::log2(commSize));
+  EXPECT_EQ(c.numPushesToFlush(), 1);
 
   // Push
   std::string s = std::to_string(commRank);
@@ -58,57 +56,33 @@ TEST(lumberjack_BinaryCommunicator, basic)
   const char* packedMessage = s.c_str();
   std::vector<const char*> receivedPackedMessages;
   c.push(packedMessage, receivedPackedMessages);
-  EXPECT_EQ(strcmp(packedMessage, origS.c_str()), 0); // Message shouldn't be
-                                                      // altered
+  EXPECT_EQ(strcmp(packedMessage, origS.c_str()), 0);
 
-  // Calculate how many children/messages you should have at this rank
-  int numChildren = 0;
-  int leftChild = (commRank*2)+1;
-  int rightChild = (commRank*2)+2;
-  if (leftChild < commSize)
+  if (commRank != 0)
   {
-    numChildren++;
+    EXPECT_EQ((int)receivedPackedMessages.size(), 0);
   }
-  if (rightChild < commSize)
+  else
   {
-    numChildren++;
-  }
-  EXPECT_EQ((int)receivedPackedMessages.size(), numChildren);
-
-  // Verify the messages received were correct
-  std::string currMessage = "";
-  bool found = false;
-
-  if (numChildren == 2)
-  {
-    currMessage = std::to_string(rightChild);
-    found = false;
-    for (auto &rm : receivedPackedMessages)
+    const int numMessagesToReceive = commSize-1;
+    EXPECT_EQ((int)receivedPackedMessages.size(), numMessagesToReceive);
+    for (int i=1 ; i<=numMessagesToReceive ; ++i)
     {
-      if (strcmp(rm, currMessage.c_str()) == 0)
+      std::string currMessage = std::to_string(i);
+      bool found = false;
+      for (auto &rm : receivedPackedMessages)
       {
-        found = true;
+        if (strcmp(rm, currMessage.c_str()) == 0)
+        {
+          found = true;
+        }
       }
-    }
-    EXPECT_EQ(found, true) << "Error: Rank: " << commRank
-                           << ": Message not received: "
-                           << currMessage << std::endl;
-  }
-
-  if (numChildren > 0)
-  {
-    currMessage = std::to_string(leftChild);
-    found = false;
-    for (auto &rm : receivedPackedMessages)
-    {
-      if (strcmp(rm, currMessage.c_str()) == 0)
+      if (!found)
       {
-        found = true;
+        std::cout << "Error: Message not received:" << currMessage << std::endl;
       }
+      EXPECT_EQ(found, true);
     }
-    EXPECT_EQ(found, true) << "Error: Rank: " << commRank
-                           << ": Message not received: "
-                           << currMessage << std::endl;
   }
 
   c.finalize();
@@ -116,7 +90,7 @@ TEST(lumberjack_BinaryCommunicator, basic)
   MPI_Barrier(MPI_COMM_WORLD);
 }
 
-TEST(lumberjack_BinaryCommunicator, pushNothing)
+TEST(lumberjack_RootCommunicator, pushNothing)
 {
   MPI_Barrier(MPI_COMM_WORLD);
 
@@ -124,7 +98,7 @@ TEST(lumberjack_BinaryCommunicator, pushNothing)
   MPI_Comm_rank(MPI_COMM_WORLD, &commRank);
 
   // Initialize
-  axom::lumberjack::BinaryTreeCommunicator c;
+  axom::lumberjack::RootCommunicator c;
   c.initialize(MPI_COMM_WORLD, 5);
   std::vector<const char*> receivedPackedMessages;
 
@@ -146,17 +120,4 @@ TEST(lumberjack_BinaryCommunicator, pushNothing)
   c.finalize();
 
   MPI_Barrier(MPI_COMM_WORLD);
-}
-
-int main(int argc, char* argv[])
-{
-  int result = 0;
-
-  ::testing::InitGoogleTest(&argc, argv);
-
-  MPI_Init(&argc, &argv);
-  result = RUN_ALL_TESTS();
-  MPI_Finalize();
-
-  return result;
 }
