+++ src/axom/lumberjack/tests/lumberjack_mpi_main.cpp
@@ -1,0 +1,25 @@
+// Copyright (c) 2017-2020, Lawrence Livermore National Security, LLC and
+// other Axom Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+
+#include "gtest/gtest.h"
+
+#include "mpi.h"
+
+#include "lumberjack_BinaryCommunicator.hpp"
+#include "lumberjack_RootCommunicator.hpp"
+
+
+int main(int argc, char* argv[])
+{
+  int result = 0;
+
+  ::testing::InitGoogleTest(&argc, argv);
+
+  MPI_Init(&argc, &argv);
+  result = RUN_ALL_TESTS();
+  MPI_Finalize();
+
+  return result;
+}
