+++ src/axom/lumberjack/tests/lumberjack_serial_main.cpp
@@ -1,0 +1,15 @@
+// Copyright (c) 2017-2020, Lawrence Livermore National Security, LLC and
+// other Axom Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+
+#include "gtest/gtest.h"
+
+#include "lumberjack_Lumberjack.hpp"
+#include "lumberjack_Message.hpp"
+#include "lumberjack_TextEqualityCombiner.hpp"
+
+int main(int argc, char** argv) {
+  ::testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
